+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increase log interval to avoid multiple Cloudwatch requests at once
 - Fix incorrect log level for terminal error code
 - Catch exceptions taking place when putLogEvents fails
+- Fix content share test video in Firefox
 
 ## [1.11.0] - 2020-06-30